Refines data cleaning and analysis for program data
Updates the data cleaning and analysis process for community college program completion data.

Specifically, it clarifies the availability of data up to 2022. It also refines the filtering of CIP codes by removing the '99' category, which includes miscellaneous programs, and adds a note to double check award levels with Zyra. Finally, updates result labels for clarity.